allow for specifying package name
