+++ pkg/cmd/get/templates/secrets.tmpl
